Add a reboot.once boolean keyword to node.conf
If once is set to false, do not remove the reboot flag before rebooting,
so that the node is ready to reboot again in the next allowed timerange.
This setup is needed to enforce a periodic reboot, with a patching script
hooked as a pre trigger for example.

If not set, or set to true, the reboot flag is removed before reboot, and
a 'nodemgr schedule reboot' is needed to rearm.

+++ lib/node.py
@@ -39,7 +39,7 @@ from rcColor import formatter
 from rcUtilities import justcall, lazy, lazy_initialized, vcall, check_privs, \
                         call, which, purge_cache_expired, read_cf, unset_lazy, \
                         drop_option
-from converters import convert_duration
+from converters import convert_duration, convert_boolean
 from comm import Crypt
 
 if sys.version_info[0] < 3:
@@ -1265,8 +1265,14 @@ class Node(Crypt):
         if statinfo.st_mode & stat.S_IWOTH:
             print("%s is world writable. abort scheduled reboot" % self.paths.reboot_flag)
             return
-        print("remove %s and reboot" % self.paths.reboot_flag)
-        os.unlink(self.paths.reboot_flag)
+        try:
+            once = self.config.get("reboot", "once")
+            once = convert_boolean(once)
+        except:
+            once = True
+        if once:
+            print("remove %s and reboot" % self.paths.reboot_flag)
+            os.unlink(self.paths.reboot_flag)
         self.reboot()
 
     def pushdisks(self):
